Fix logic related to cleaning up old media files (#2153)
* Fix logic related to cleaning up old media files

- we had logic to cleanup old processed media files.
- this logic was broken - since the media file names were random uuids.
- and the logic was checking if the uuid matched a post id.
- we need to cleanup this whole logic separately - should keep this in mind when moving over feedData to core.
- until then, i set the media file names based on the contentID (could be msgId, postId, commentId or linkPreviewId) and the mediaIndex.
- we do this now in both chatData and feedData.
- This way, when cleaning up these old files - we wont accidently delete files where the post is not fully uploaded yet.
- this is what happened with pooja's post the other day.
- tested multiple times and this works cleanly.

+++ Halloapp/Network/ImageServer.swift
@@ -591,10 +591,19 @@ class ImageServer {
         let encryptedMediaFileURL = processedMediaFileURL.appendingPathExtension(encryptedSuffix)
 
         do {
-            try FileManager.default.removeItem(at: originalMediaFileURL)    // xxx.jpg
-            try FileManager.default.removeItem(at: encryptedMediaFileURL)   // xxx.processed.jpg.enc
+            if FileManager.default.fileExists(atPath: originalMediaFileURL.absoluteString) {
+                try FileManager.default.removeItem(at: originalMediaFileURL)    // xxx.jpg
+                DDLogDebug("ImageServer/cleanUpUploadData/success: \(originalMediaFileURL)")
+            }
+        } catch {
+            DDLogError("ImageServer/cleanUpUploadData/error [\(error)]")
         }
-        catch {
+        do {
+            if FileManager.default.fileExists(atPath: originalMediaFileURL.absoluteString) {
+                try FileManager.default.removeItem(at: encryptedMediaFileURL)   // xxx.processed.jpg.enc
+                DDLogDebug("ImageServer/cleanUpUploadData/success: \(encryptedMediaFileURL)")
+            }
+        } catch {
             DDLogError("ImageServer/cleanUpUploadData/error [\(error)]")
         }
     }
